List goal validation.

+++ open-life-server/V1/Goals/ListGoals/ListGoalController.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -10,56 +11,90 @@ namespace open_life_server.V1.Goals.ListGoals
     public class ListGoalController : ControllerBase
     {
         private readonly GoalsContext _context;
+        private readonly IListGoalValidator _validator;
 
-        public ListGoalController(GoalsContext context)
+        public ListGoalController(GoalsContext context, IListGoalValidator validator)
         {
             _context = context;
+            _validator = validator;
         }
 
         // GET: api/ListGoal
         [HttpGet]
-        public IEnumerable<ListGoal> Get()
+        [ProducesResponseType(typeof(IEnumerable<ListGoal>), StatusCodes.Status200OK)]
+        public IActionResult Get()
         {
-            return _context.ListGoals.Include(g => g.Items).ToList();
+            return Ok(_context.ListGoals.Include(g => g.Items).ToList());
         }
 
         // GET: api/ListGoal/5
         [HttpGet("{id}")]
-        public ListGoal Get(int id)
+        [ProducesResponseType(typeof(ListGoal), StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        public IActionResult Get(int id)
         {
-            return _context.ListGoals.Include(g => g.Items).Single(g => g.ListGoalId == id);
+            var goal = _context.ListGoals.Include(g => g.Items).Single(g => g.ListGoalId == id);
+
+            if (goal == null)
+                return NotFound();
+
+            return Ok(goal);
         }
 
         // POST: api/ListGoal
         [HttpPost]
-        public void Post([FromBody] ListGoal value)
+        [ProducesResponseType(typeof(ListGoal), StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        public IActionResult Post([FromBody] ListGoal value)
         {
-            _context.ListGoals.Add(value);
+            if (!_validator.Valid(value))
+                return BadRequest(_validator.GetInvalidMessage(value));
+
+            var goal = _context.ListGoals.Add(value).Entity;
             _context.SaveChanges();
+
+            return Ok(goal);
         }
 
         // PUT: api/ListGoal/5
         [HttpPut("{id}")]
-        public void Put(int id, [FromBody] ListGoal value)
+        [ProducesResponseType(typeof(ListGoal), StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        public IActionResult Put(int id, [FromBody] ListGoal value)
         {
+            if (!_validator.Valid(value))
+                return BadRequest(_validator.GetInvalidMessage(value));
+
             var listToUpdate = _context.ListGoals.Find(id);
+            if (listToUpdate == null)
+                return NotFound();
 
             listToUpdate.Name = value.Name;
             listToUpdate.ListName = value.ListName;
             listToUpdate.Target = value.Target;
             listToUpdate.Items = value.Items;
 
-            _context.ListGoals.Update(listToUpdate);
+            var goal = _context.ListGoals.Update(listToUpdate).Entity;
             _context.SaveChanges();
+
+            return Ok(goal);
         }
 
         // DELETE: api/ListGoal/5
         [HttpDelete("{id}")]
-        public void Delete(int id)
+        [ProducesResponseType(typeof(ListGoal), StatusCodes.Status202Accepted)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        public IActionResult Delete(int id)
         {
             var listToDelete = _context.ListGoals.Find(id);
+            if (listToDelete == null)
+                return NotFound();
+
             _context.ListGoals.Remove(listToDelete);
             _context.SaveChanges();
+
+            return Accepted();
         }
     }
 }
